+++ NEWS.md
@@ -1,3 +1,30 @@
+# posterior 1.6.0
+
+### Enhancements
+
+* Add `exclude` option to `subset_draws()`, which can be used to exclude
+  the matched selection.
+* Add `are_log_weights` option to `pareto_smooth()`, which is necessary
+  for correct Pareto smoothing computation if the input vector
+  consists of log weights.
+* Add `pareto_smooth` option to `weight_draws()`, to Pareto smooth
+  weights before adding to a draws object.
+* Add individual Pareto diagnostic functions (`pareto_khat()`,
+  `pareto_khat_threshold()`, `pareto_min_ss()`, `pareto_convergence_rate()`)
+* `thin_draws()` now automatically thins draws based on ESS by default,
+  and non-integer thinning is possible.
+* Matrix multiplication of `rvar`s can now be done with the base matrix 
+  multiplication operator (`%*%`) instead of `%**%` in R >= 4.3.
+* `variables()`, `variables<-()`, `set_variables()`, and `nvariables()` now
+  support a `with_indices` argument, which determines whether variable names
+  are retrieved/set with (`"x[1]"`, `"x[2]"` ...) or without (`"x"`) indices
+  (#208).
+* Add `extract_variable_array()` function to extract variables with indices
+  into arrays of iterations x chains x any remaining dimensions (#340).
+* For types that support `factor` variables (`draws_df`, `draws_list`, and 
+  `draws_rvars`), `extract_variable()` and `extract_variable_matrix()` can 
+  now return `factor`s.
+
 # posterior 1.5.0
 
 ### Enhancements